Test refactor complete
